feat: add nextjs, delete react-router & vite, migrate to next app router
