+++ .gitignore
@@ -11,6 +11,7 @@ node_modules
 dist
 dist-ssr
 coverage
+.next
 *.local
 
 # Editor directories and files
