+++ src/app/QueryProvider.tsx
@@ -1,0 +1,17 @@
+'use client';
+
+import type { ReactNode } from 'react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import ErrorBoundary from '../components/ErrorBoundary/ErrorBoundary';
+
+const queryClient = new QueryClient();
+
+function QueryProvider({ children }: { children: ReactNode }) {
+  return (
+    <ErrorBoundary fallback={<p>UI Error has happened</p>}>
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    </ErrorBoundary>
+  );
+}
+
+export default QueryProvider;
