--- src/main.tsx
@@ -1,24 +1,0 @@
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import './index.css';
-import App from './app/App.tsx';
-import ErrorBoundary from './components/ErrorBoundary/ErrorBoundary.tsx';
-import { BrowserRouter } from 'react-router-dom';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
-const root = document.querySelector('#root');
-const queryClient = new QueryClient();
-
-if (root) {
-  createRoot(root).render(
-    <StrictMode>
-      <ErrorBoundary fallback={<p>UI Error has happened</p>}>
-        <BrowserRouter>
-          <QueryClientProvider client={queryClient}>
-            <App />
-          </QueryClientProvider>
-        </BrowserRouter>
-      </ErrorBoundary>
-    </StrictMode>
-  );
-}
